Move read-only security module to plugin toolkit
