+++ presto-plugin-toolkit/src/main/java/com/facebook/presto/plugin/base/security/ReadOnlySecurityModule.java
@@ -11,7 +11,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.facebook.presto.hive;
+package com.facebook.presto.plugin.base.security;
 
 import com.facebook.presto.spi.connector.ConnectorAccessControl;
 import com.google.inject.Binder;
